feat: read saved excalidraw data
