+++ package.json
@@ -57,7 +57,7 @@
     "@deckdeckgo/offline": "^0.0.6",
     "@deckdeckgo/social-img": "^1.3.0",
     "@deckdeckgo/studio": "^0.0.21",
-    "@deckdeckgo/sync": "^0.0.28",
+    "@deckdeckgo/sync": "file:../../deckdeckgo/deckdeckgo/providers/sync",
     "@papyrs/stylo": "^0.0.28",
     "@papyrs/what-is-internet-identity": "^0.0.2",
     "@stencil/core": "^2.15.2",
